+LazySnapping fixed but needs improvement +minor UI update

+++ src/maskeditor/segmentation/lazysnapping/maxflow/graph.h
@@ -66,9 +66,11 @@ public:
 
 	/* Type of edge weights.
 	   Can be changed to char, int, float, double, ... */
-	typedef short captype;
+	//typedef short captype;
+    typedef float captype;
 	/* Type of total flow */
-	typedef int flowtype;
+	//typedef int flowtype;
+    typedef float flowtype;
 
 	typedef void * node_id;
